clova-service/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.25.0
donut-python>=1.0.9
craft-text-detector>=0.4.3
opencv-python>=4.6.0
Pillow>=9.0.0
numpy>=1.21.0
python-multipart>=0.0.6
scikit-image>=0.19.0
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.0.0